requirements.txt
python-telegram-bot==20.6
deep-translator==1.11.4
